Post-merge PHP cleanup (no HTML)

+++ php/Parser/File.php
@@ -21,8 +21,7 @@ final class File extends Parser
 
     protected function getTotalScore(): int
     {
-        // @phpstan-ignore return.type
-        return array_sum(array_map(array_sum(...), $this->model));
+        return (int) array_sum(array_map(array_sum(...), $this->model));
     }
 
     protected function getScore(string $featureKey, string $sequence): int
